requirements.txt
openai
streamlit
faiss-cpu 
tiktoken
python-dotenv